{{ cookiecutter.project_name }}/requirements/common.txt
aerich>=0.7.1
asyncpg>=0.27.0
esmerald>=0.2.5
gunicorn>=20.1.0
httpx>=0.23.0
postgres>=4.0.0
psycopg[pool]>=3.1.4
redis>=3.5.3
sib_api_v3_sdk>=7.5.0
tortoise-orm>=0.19.2
uvicorn>=0.19.0
